requirements.txt
absl-py
astunparse
attrs
certifi
charset-normalizer
colorama
contourpy
cycler
flatbuffers
fonttools
future
gast
google-pasta
grpcio
h5py
idna
joblib
keras
kiwisolver
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml_dtypes
namex
numpy
opt_einsum
optree
packaging
pillow
protobuf
Pygments
pyparsing
pyswarms
python-dateutil
PyYAML
requests
rich
scikit-learn
scipy
setuptools
six
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
tqdm
typing_extensions
urllib3
Werkzeug
wheel
wrapt
